requirements-basic.txt
# Brazilian E-commerce Data Pipeline - Basic Dependencies (Python 3.12+ Compatible)
# =================================================================================
# This version excludes Apache Beam due to Python version compatibility issues
# Use this for MongoDB import and basic data processing functionality

# Core Data Processing
pandas>=1.3.0,<2.0.0
numpy>=1.21.0,<2.0.0

# MongoDB
pymongo>=4.0.0,<5.0.0

# Environment Variables
python-dotenv>=0.19.0,<2.0.0

# Google Cloud Storage (direct, without Apache Beam)
google-cloud-storage>=2.0.0,<3.0.0

# Additional Utilities
typing-extensions>=4.0.0

# Jupyter Notebook support
jupyter>=1.0.0,<2.0.0
ipykernel>=6.0.0,<7.0.0

# Data Visualization
matplotlib>=3.5.0,<4.0.0
seaborn>=0.11.0,<1.0.0
plotly>=5.0.0,<6.0.0

# CSV processing utilities
csvkit>=1.0.0,<2.0.0
